Use parseEnvelope from @sentery/utils
