+++ package.json
@@ -57,18 +57,18 @@
     "e2e": "cross-env TS_NODE_PROJECT=tsconfig.json xvfb-maybe mocha --require ts-node/register/transpile-only --retries 3 ./test/e2e/*.ts"
   },
   "dependencies": {
-    "@sentry/browser": "7.24.1",
-    "@sentry/core": "7.24.1",
-    "@sentry/node": "7.24.1",
-    "@sentry/types": "7.24.1",
-    "@sentry/utils": "7.24.1",
+    "@sentry/browser": "7.28.1",
+    "@sentry/core": "7.28.0",
+    "@sentry/node": "7.28.0",
+    "@sentry/types": "7.28.0",
+    "@sentry/utils": "7.28.0",
     "deepmerge": "4.2.2",
     "tslib": "^2.3.1"
   },
   "devDependencies": {
-    "@sentry-internal/eslint-config-sdk": "7.24.1",
-    "@sentry-internal/typescript": "7.24.1",
-    "@sentry/tracing": "7.24.1",
+    "@sentry-internal/eslint-config-sdk": "7.28.0",
+    "@sentry-internal/typescript": "7.28.0",
+    "@sentry/tracing": "7.28.0",
     "@types/busboy": "^0.2.3",
     "@types/chai": "^4.2.10",
     "@types/chai-as-promised": "^7.1.5",
